# apps/api/internal/presentations/store.go
package presentations

import (
	"context"
	"encoding/json"
	"errors"
	"time"
)

var (
	ErrNotFound        = errors.New("presentation not found")
	ErrInvalidPosition = errors.New("invalid slide position")
	ErrSlideHasResults = errors.New("slide has live results")
)

type Slide struct {
	ID       string          `json:"id"`
	Position int             `json:"position"`
	Kind     string          `json:"kind"`
	Content  json.RawMessage `json:"content"`
}
type Presentation struct {
	ID        string          `json:"id"`
	Title     string          `json:"title"`
	Settings  json.RawMessage `json:"settings"`
	CreatedAt time.Time       `json:"created_at"`
	UpdatedAt time.Time       `json:"updated_at"`
	Slides    []Slide         `json:"slides"`
}
type PresentationSummary struct {
	ID               string          `json:"id"`
	Title            string          `json:"title"`
	Settings         json.RawMessage `json:"settings"`
	SlideCount       int             `json:"slide_count"`
	ParticipantCount int             `json:"participant_count"`
	CreatedAt        time.Time       `json:"created_at"`
	UpdatedAt        time.Time       `json:"updated_at"`
}
type PresentationPatch struct {
	Title    *string
	Settings json.RawMessage
}
type SessionLocator struct {
	SessionID      string `json:"session_id"`
	PresentationID string `json:"presentation_id"`
}
type QuestionResultCursor struct {
	Score       int       `json:"s"`
	SubmittedAt time.Time `json:"t"`
	AnswerID    string    `json:"i"`
}
type QuestionResultsQuery struct {
	Limit  int
	Cursor *QuestionResultCursor
}
type QuestionOptionResult struct {
	OptionIndex     int    `json:"option_index"`
	OptionID        string `json:"option_id"`
	OptionText      string `json:"option_text"`
	IsCorrect       bool   `json:"is_correct"`
	NumberOfSubmits int    `json:"number_of_submits"`
}
type QuestionLeaderboardEntry struct {
	ParticipantID string    `json:"participant_id"`
	DisplayName   string    `json:"display_name"`
	Avatar        string    `json:"avatar,omitempty"`
	Score         int       `json:"score"`
	Rank          int       `json:"rank"`
	TimeTakenMS   *int64    `json:"time_taken_ms"`
	SubmittedAt   time.Time `json:"submitted_at"`
	answerID      string
}
type QuestionResultsPage struct {
	SessionID       string                     `json:"session_id"`
	QuestionSlideID string                     `json:"question_slide_id"`
	ResponseCount   int                        `json:"response_count"`
	Options         []QuestionOptionResult     `json:"options"`
	Leaderboard     []QuestionLeaderboardEntry `json:"leaderboard"`
	Limit           int                        `json:"limit"`
	HasMore         bool                       `json:"has_more"`
	NextCursor      *string                    `json:"next_cursor,omitempty"`
}
type Store interface {
	ListOwned(context.Context, string) ([]PresentationSummary, error)
	FindOwned(context.Context, string, string) (Presentation, error)
	Create(context.Context, string, string, json.RawMessage) (Presentation, error)
	Update(context.Context, string, string, PresentationPatch) (Presentation, error)
	Delete(context.Context, string, string) error
	Duplicate(context.Context, string, string, string) (Presentation, error)
	LatestSession(context.Context, string, string) (SessionLocator, error)
	DeleteResults(context.Context, string, string) error
	QuestionResults(context.Context, string, string, string, string, QuestionResultsQuery) (QuestionResultsPage, error)
	CreateSlide(context.Context, string, string, int, string, json.RawMessage) (Slide, error)
	ReplaceSlide(context.Context, string, string, string, int, string, json.RawMessage) (Slide, error)
	DeleteSlide(context.Context, string, string, string) error
	ReorderSlides(context.Context, string, string, []string) error
}