# apps/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ProSlides API
  version: 0.1.0
  description: Contract-first API for the Go modular monolith.
servers:
  - url: /
paths:
...
  /api/v1/presentations/{presentationId}:
    get:
      operationId: getPresentation
      summary: Return one presentation and its ordered slides for its owner.
      description: Requires the current authenticated session. Returns 404 when the presentation does not exist or is not owned by the current account, so identifiers are not disclosed.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
      responses:
        '200':
          description: The presentation with slides ordered by ascending position.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Presentation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updatePresentation
      summary: Update an owned presentation's title or visual settings.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdatePresentationRequest' } } }
      responses:
        '200': { description: Updated presentation., content: { application/json: { schema: { $ref: '#/components/schemas/Presentation' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deletePresentation
      summary: Permanently delete an owned presentation and its dependent data.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204': { description: Presentation deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/presentations:
    get:
      operationId: listPresentations
      summary: List the current owner's presentations in stable recent-first order.
      responses:
        '200':
          description: Bounded owner presentation list.
          content: { application/json: { schema: { type: array, maxItems: 200, items: { $ref: '#/components/schemas/PresentationSummary' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createPresentation
      summary: Create an empty presentation owned by the current account.
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePresentationRequest' }
      responses:
        '201':
          description: Presentation created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Presentation' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
  /api/v1/presentations/{presentationId}/duplicate:
    post:
      operationId: duplicatePresentation
      summary: Atomically duplicate an owned presentation and all ordered slides.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreatePresentationRequest' } } }
      responses:
        '201': { description: Duplicated presentation., content: { application/json: { schema: { $ref: '#/components/schemas/Presentation' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/presentations/{presentationId}/latest-session:
    get:
      operationId: getLatestPresentationSession
      summary: Resolve the newest live session for an owned presentation.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
      responses:
        '200': { description: Latest session locator., content: { application/json: { schema: { $ref: '#/components/schemas/LiveSessionLocator' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/presentations/{presentationId}/sessions/{sessionId}/questions/{slideId}/results:
    get:
      operationId: getQuestionResults
      summary: Return persisted option counts and a stable bounded per-question leaderboard for an owned session.
      description: This is the Go-native replacement for Django's Rust result-ingestion endpoints. Results are derived from durable answers and never accept a second externally supplied score ledger.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - $ref: '#/components/parameters/SessionId'
        - { name: slideId, in: path, required: true, schema: { type: string, format: uuid } }
        - $ref: '#/components/parameters/RosterLimit'
        - $ref: '#/components/parameters/RosterCursor'
      responses:
        '200':
          description: Question results for the requested owned live session.
          content: { application/json: { schema: { $ref: '#/components/schemas/QuestionResultsPage' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/presentations/{presentationId}/results:
    delete:
      operationId: deletePresentationResults
      summary: Delete all live sessions and results for an owned presentation while preserving its content.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204': { description: Results deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/presentations/{presentationId}/slides:
    post:
      operationId: createSlide
      summary: Add a slide to an owned presentation.
      description: Requires an authenticated owner and a CSRF token. Inserting at an occupied zero-based position atomically shifts following slides. Returns 404 for absent or non-owned presentations.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSlideRequest' }
      responses:
        '201':
          description: Slide created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Slide' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/presentations/{presentationId}/slides/reorder:
    post:
      operationId: reorderSlides
      summary: Atomically replace the complete slide order for an owned presentation.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ReorderSlidesRequest' } } }
      responses:
        '204': { description: Slide positions updated. }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/presentations/{presentationId}/slides/{slideId}:
    put:
      operationId: replaceSlide
      summary: Replace the editable definition of one owned slide.
      description: Moving to an occupied zero-based position atomically compacts the other positions. A slide with durable answers may be moved, but its question definition cannot change until presentation results are reset.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - { name: slideId, in: path, required: true, schema: { type: string, format: uuid } }
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateSlideRequest' } } }
      responses:
        '200': { description: Replaced slide., content: { application/json: { schema: { $ref: '#/components/schemas/Slide' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: The requested edit would change a slide with durable live answers. }
    delete:
      operationId: deleteSlide
      summary: Delete one owned slide and compact following positions.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - { name: slideId, in: path, required: true, schema: { type: string, format: uuid } }
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204': { description: Slide deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: The slide has durable live answers; reset presentation results before deleting it. }
  /api/v1/presentations/{presentationId}/questions:
    post:
      operationId: createQuestionSlide
      summary: Add a single- or multiple-choice question slide.
      parameters:
        - $ref: '#/components/parameters/PresentationId'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateQuestionRequest' }
      responses:
        '201': { description: Question slide created. }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CSRFError' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RosterLimit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    RosterCursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned by the preceding page.
      schema: { type: string, maxLength: 512 }
...
    PresentationId:
      name: presentationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Presentation:
      type: object
      required: [id, title, settings, created_at, updated_at, slides]
      properties:
        id: { type: string, format: uuid }
        title: { type: string, minLength: 1, maxLength: 500 }
        settings: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        slides:
          type: array
          items: { $ref: '#/components/schemas/Slide' }
    PresentationSummary:
      type: object
      required: [id, title, settings, slide_count, participant_count, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        settings: { type: object, additionalProperties: true }
        slide_count: { type: integer, minimum: 0 }
        participant_count: { type: integer, minimum: 0 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreatePresentationRequest:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 500 }
        settings: { type: object, additionalProperties: true }
    UpdatePresentationRequest:
      type: object
      minProperties: 1
      properties:
        title: { type: string, minLength: 1, maxLength: 500 }
        settings: { type: object, additionalProperties: true }
    CreateSlideRequest:
      type: object
      required: [position, kind, content]
      properties:
        position: { type: integer, minimum: 0 }
        kind: { type: string, minLength: 1, maxLength: 100 }
        content: { type: object, additionalProperties: true }
    ReorderSlidesRequest:
      type: object
      required: [slide_ids]
      properties:
        slide_ids:
          type: array
          maxItems: 1000
          uniqueItems: true
          items: { type: string, format: uuid }
    CreateQuestionRequest:
      type: object
      required: [position, text, question_type, options]
      properties:
        position: { type: integer, minimum: 0 }
        text: { type: string, minLength: 1 }
        question_type: { type: string, enum: [single, multiple] }
        question_time: { type: integer, minimum: 1, maximum: 86400, default: 30 }
        max_point: { type: integer, minimum: 1, default: 100 }
        min_point: { type: integer, minimum: 0, default: 0 }
        faster_answers_more_points: { type: boolean, default: false }
        partial_scoring:
          type: boolean
          default: false
          description: When true, score is max(0, correct selections minus incorrect selections) divided by the number of correct options; the scoring policy is an implementation boundary and may be replaced later.
        options:
          type: array
          minItems: 2
          items:
            type: object
            required: [text, is_correct]
            properties:
              text: { type: string, minLength: 1 }
              is_correct: { type: boolean }
...
    LiveSessionLocator:
      type: object
      required: [session_id, presentation_id]
      additionalProperties: false
      properties:
        session_id: { type: string, format: uuid }
        presentation_id: { type: string, format: uuid }
...
    QuestionOptionResult:
      type: object
      required: [option_index, option_id, option_text, is_correct, number_of_submits]
      additionalProperties: false
      properties:
        option_index: { type: integer, minimum: 0 }
        option_id: { type: string }
        option_text: { type: string }
        is_correct: { type: boolean }
        number_of_submits: { type: integer, minimum: 0 }
    QuestionLeaderboardEntry:
      type: object
      required: [participant_id, display_name, score, rank, submitted_at]
      additionalProperties: false
      properties:
        participant_id: { type: string, format: uuid }
        display_name: { type: string }
        avatar: { type: string }
        score: { type: integer, minimum: 0 }
        rank: { type: integer, minimum: 1 }
        time_taken_ms: { type: integer, minimum: 0, nullable: true }
        submitted_at: { type: string, format: date-time }
    QuestionResultsPage:
      type: object
      required: [session_id, question_slide_id, response_count, options, leaderboard, limit, has_more]
      additionalProperties: false
      properties:
        session_id: { type: string, format: uuid }
        question_slide_id: { type: string, format: uuid }
        response_count: { type: integer, minimum: 0 }
        options:
          type: array
          items: { $ref: '#/components/schemas/QuestionOptionResult' }
        leaderboard:
          type: array
          items: { $ref: '#/components/schemas/QuestionLeaderboardEntry' }
        limit: { type: integer, minimum: 1, maximum: 100 }
        has_more: { type: boolean }
        next_cursor: { type: string, nullable: true }
...
    Slide:
      type: object
      required: [id, position, kind, content]
      properties:
        id: { type: string, format: uuid }
        position: { type: integer, minimum: 0 }
        kind: { type: string, minLength: 1, maxLength: 100 }
        content:
          type: object
          additionalProperties: true
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: No valid current session exists.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    CSRFError:
      description: The CSRF header and cookie do not match a valid current session.
      content:
        application/json:
              schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: The requested resource does not exist or is not visible to the current account.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...